requirements.txt
# Instructor evaluation system dependencies
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
pandas==2.1.3
playwright==1.40.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
openai==1.3.5
anthropic==0.7.1
pydantic==2.5.0
aiohttp==3.9.1
PyGithub==2.1.1
